[HALON-374] Prepare process stop rules for changes
In preparation of implementing different ways to effectively stop a
process (i.e. not just `systemctl`), making the process stop and node
stop rules a bit saner was in order. Similarly as in HALON-373, we get
rid of the notion of working on multiple processes. Instead, we use a
job to work on one process at a time. A node rule is responsible for
managing the set of these jobs. The node rule now simply dispatches the
process stop requests and waits for results. A few message types were
changed to eliminate completely redundant look-ups and failure handlers.

The way notifications are handled has been changed. Process stop rule
notified quite a bit. The problem with the way notifications were being
done is that we were always giving something up: configurable timeout,
sane interface, nice state update, …. For example, consider wanting to
wait for `PSFailed str` with some dynamically determined timeout:
`mkPhaseNotify` is out of question, timeout had to be specified on RuleM
level so we can't leak it out of PhaseM. The same applies to `str`: we
have to know ahead of time what state we're waiting for exactly or we
had to write a predicate accepting all states or figure out all the
possible values.

There was however already a solution for a similar issue: dispatcher.
Refer to Haddock for use example. In hindsight, it may be useful to
allow multiple dispatchers in a single rule. There still remain a few
‘ugly’ notification handlers: these are the handlers that deal
with *all* objects. As how they work is fairly simple and they weren't
directly in process start/stop rules, they were left to be but they
should be ported to a similar interface Soon™. All the notification
helpers were moved to their own module.