HALON-570: fix possible possible ignorance of the new drive. (#1130)
* HALON-570: fix possible possible ignorance of the new drive.

In HALON-570 issued Smart request for the old drive instead
of a new drive, because drive candidate that was created in
HPI rule was never actualized. This happened because HPI
rule incorrectly decided that drive was powered on instead
of beign inserted.

Currently sspl:monitor-status-hpi rule doesn't check if drive
was changed or not, when decide what happen with disk.
That may lead to a problems when halon could make incorrect
decisions. The problematic case is:
0. Drive fails, marked as removed, etc.
1. SSPL-LL sends HPI event for some reason -> this leads to the
   drop of the DriveRemoved status
2. SSPL-LL sends HPI event about new drive, beceuse there is
   no drive removed status, halon decide that it's a repower.

The fix is to make halon aware if it was a new drive or not.
It seems it's not easy to fix `1.` as HPI message is the only
source for drive insertion event, this means that if
monitor-drivemanager and hpi messages will be unordered, then
halon will miss drive insertion.

* HALON-570: do not loose StorageDevice links when updating.

* Mark device as present when it's a new drive.

Previously we did not do this.

* Remove device identifiers from old disk.

Identifiers get added to the new disk by `attachStorageDeviceReplacement`,
so should be removed from the old disk before merge.

* Add device path when not present.

Previously wasn't added in this branch, meaning replacement
drive didn't have the right path.

* HALON-570: set correct path on drive-manager request.

When drive-manager request arrive halon find disk in the slot.
Previously halon attached path information to that drive, but
instead halon should check if there is a replacement for that
drives and attach path there.

When HPI event about new drive arrives HALON should not insert
id in enclosure field to candidate, otherwise drive-manager rule
may decide that disk is already inserted in the slot.

attachStorageDeviceReplacement can be called multiple times,
in such case no new replacement drive is created, instead we
just attach new identifier to that device.

* HALON-570: update confd only when disk path is known.

* Don't test that new SN is the same as old SN.