[HALON-574] Async controlProcess commands
Spawn systemctl commands on nodes asynchronously. This means processes
on same boot-level no longer have to wait on each other. This should
speed up concurrent process start/stop (bootstrap/teardown) and greatly
reduce the risk of bogus process rule timeouts.

We check the commands being ran for uniqueness: if we're handling a
certain request already, we won't do it again if the process for that
command is still running. We keep a map of running processes. Note that
nothing stops us here from running stop and start commands at once for
example: the user has to make sure that the issued control commands are
sane.

In the near future, perhaps an "abort" control command is in order: the
rules could request that we kill the running command as part of
clean-up. This means we won't have late message results coming (modulo a
small race of abort command being sent and actually processed which
should be OK: rule can account for this) and we can be assured that if
we're requesting process start, our environment is somewhat sane if
process stop rule has previously failed.