[HALON-580] Mero mock service implementation
The main idea is implementation of a `halon:m0d` service replacement
that's used during testing. RC relies on the service on communication
with mero, controlling processes on nodes, keepalive and notification
delivery. Many rules are written to rely on these things to properly
function: it is of no fault of the rules; if the requirement is that
mero has to have received the notification then we can not change the
rule not to do so.

Instead, a new implementation is needed. RC is tricked into using it by
storing the `Service MeroConf` value to be used in RG. During normal
operation this is, by default, `HA.Services.Mero.m0d_real`. During
tests, this value gets replaced. Beyond this lookup RC itself is
oblivious which means we don't have to change any rules and everything
can be tested as-is.

The mock service does not run mero. To compensate, it simply uses a
local state (`MockLocalState`) to keep track of the information it is
given. One caveat is that the service needs additional initialisation
before it can run as it has nowhere else to get that information from.
Refer to `testRules` in `HA.Castor.Story.Tests` for actual
implementation but the gist is that the mock service will not report
that it has started until additional information is sent to it. The
information required is about processes and services on the node this
service is running on. This information is used to send out
`ProcessEvent` and `ServiceEvent` messages to RC; again, mero is not
there to do that for us but rules rely on it. In practice we don't start
any processes but simply generate PIDs sequentially. These PIDs are also
stored; this is needed for `ProcessEvent` information. Currently RC does
not interact with processes directly (i.e. it always goes through
`halon:m0d`) so this method is sufficient.

Halon allows us to start services on specific `NodeId`s. In fact,
`runTests` will generate nodes for us if requested. The above
implementation which does not use any physical resources (notably, no
kernel module is loaded, no processes are started, no files are written)
which means we can start multiple services, one on each node. In fact,
if we couple this with a unique hostname for each of the nodes, we're
able to properly start and join multiple nodes to cluster on a single
physical machine. See `testClusterBootstrap` test for an example. As
service initialisation &c. logic is inside test runner itself, the test
simply tells the runner to provide a bootstrapped cluster with 4 nodes
then exits once the environment is ready.

Beyond service initialisation hack, we really *are* just asking the
cluster to start as it would on real cluster. This means initial data
used for tests had to be updated. Initially not even CST_HA process was
present: it's needed for the mock. We also needed some processes for
expander reset. Process tests were deleted and instead replaced with a
stop then start on boot level 1 process. To accomodate all this, we now
supply every generated host with a halon process, confd, mds, ios and
m0t1fs. To know how many nodes we require and with what hostnames, we
now generate initial data at start, generate `TestOptions` from that and
feed them into test runner. This means we simply customise initial data
used in test if we require a custom setup and everything should work
automatically. As bootstrapping the (in-memory) cluster is not necessary
for every test and takes time, that's configurable too: test runner can
be asked to only start halon:m0d services or to just leave cluster in
default state. This means we only incur time ‘penalty’ on some tests.

Most other changes were to adapt existing tests to a multi-node
scenario. Further, as we now always generate nodes (i.e. in case user
requests 1 node for tests, test runner generates 2 nodes: one runs test
and one is provided for use. The un-provided test is RC node.),
`getSelfNode` and similar no longer do the expected thing. `getHostName`
and similar also no longer did the expected thing. Basically anything
that assumed single-node scenario with co-location with RC broke. Any
direct notification use in tests was removed: this was extremely fragile
and broke frequently. It also just didn't test the right thing: often
some old notifications would be accepted. `halon:m0d` mock service now
automatically acks all notifications (but can be configured to fail, see
`_m_not_notifiable` and `SetNotNotifiable` though no existing test
relied on such functionality; new tests can exercise rules through
notification failure paths). Where notifications were used to determine
that an object entered some state, `waitState` helper is used instead
which queries RG through MM directly until it sees the object
transition (though note it's not a perfect solution: fast (< retry
interval of `waitState`) state changes have potential to go unnoticed;
it's not an issue in existing tests however). Similiarly, other data is
queried from RG instead of hard-coded or wrong. Test runner no longer
sets up notifications that tests rely on itself. Instead, tests should
do it. For notifications used by runner itself, they are only subscribed
to while they are needed (`withSubscription`).

There are a few things to note. This PR does not re-organise tests or
write many new ones (in fact only process tests were replaced and the
‘trivial’ bootstrap test was added). Many should be rewritten in terms
of waiting for rule results and grouped. Multi-node bootstrap on VM is
possible but it turns out halon can't handle too many nodes: the tests
uses 4 nodes. I was able to bootstrap once with 10 nodes. Seems for more
than 5 nodes or so, halon starts to really struggle. Examining eventlog
puts heavy blame on `eq::record::event` in EQ, `statusProcess` and
`sendProcessEvents` in `halon:m0d`. It's easy to see why this is the
case: the service uses `promulgateWait` which blocks and also means
notifications don't get acked that fast. We can relieve this with
`promulgateEQ` but: it means it doesn't model real `halon:m0d`
accurately (which uses `promulgateWait`) but also does not actually
help; `eq::record::event` works very hard anyway to replicate each
messages. There are many, many messages. Let's say we have a 20 node
setup. Trying to run cluster bootstrap on such a setup showed 1700
messages notification sets being received (so `1700 / 20 = 85`
notifications about state changes which sounds reasonable) before the
test timed out. 85 notifications * 20 nodes (20 `halon:m0d` services) *
3 recipients on each node gives about 5000 replicated messages. The test
timeout was 120 seconds; let's ignore 20 seconds on initial service
initialisation and other test setup; this gives about 500
replicated (and `promulgateWait`ed but this doesn't matter) messages per
second. Result is everything times out &c. This is not a problem with
the mock but with how we process notifications in general however it
explains why we can't do big bootstraps (yet). Last problem I would like
to mention is that even though we start services on each node, RMQ is a
special case and only a single proxy instance is opened: I have learned
that it's very picky and don't want to mess with it by opening multiple
connections unless it becomes necessary in the future. I have not tried
yet.

In case of 1 node bootstrap seems to take some 30-40 seconds on my
machine. It should be investigated why. Some is lost to test setup but
some seems to be slow halon. Test setup should be sped up too.

As a closing note, I wonder if this allows us to run USE_MERO tests (and
bootstrap!) outside of the dev VM.