[HALON-535] Don't run SSPL RaidAdd concurrently (#1095)
It could happen that both replace and failure rules would attempt to add
the same device to an array concurrently. While removing the command
from one rule and not the other would be difficult (to test mostly:
there are multiple code paths and entry points), we can solve it by
introducing a job which does the insertion.

This way both rules are free to run concurrently without any issue or
bad interleaving.