[HALON-590] Give leveldb a chance to finish writes
We see a segfault on write in leveldb with drive-manager-works test. The
test is sensitive to changes: if we add any delays or change some
messages, we stop failing. I suspect it's because something (trim?)
keeps writing to the db; meanwhile the test finished and `unsafeClose`
is called on the db handle. If this is the case then it's unclear how to
fix this properly: perhaps we need to set some value in replicated-log
when writes are happening and block db close on until after writes are
finished.

In meantime, just add a small timeout to unblock master.