[HALON-583] Increase timeouts related to process stop
Increase timeout for systemctl stop command. This impacts at what rate
will timing out processes be reported back to cluster stop rule so
increase a barrier timeout too; for the same reason increase how long
halonctl should wait for cluster stop to happen: this value is already
technically too low because we have multiple bootlevels but it will have
to do.

A future task should be the removal of barrier timeout all together:
it's not necessary if process stop rule always returns *and* we don't
want to ‘skip ahead’ in cluster teardown if some processes are slow.