Prettify the text log. (#1140)
The decision-log service now collects all logs for a given
location, much as it used to. It also now outputs the underlying
events, with the exception of context events which are collated
together by location. A new pretty(ish) printer is written to
display these logs in a nice fashion.

The binary printer has been changed to output the JSON encoding
rather than using `Binary`. This is primarily to avoid pushing
`SafeCopy` down into CEP, where it should not be needed. We also
don't need Binary upgrade capability for the logs; indeed, the
ability to parse them in other languages should make the JSON
output more useful.