Add code for CAS/DIX into Halon.
- Update confd to support changes to Mero code.
- Add support for imeta_pver into filesystem object.
- Create imeta pool and pver on initial data load.

There is currently a bug in this - if there are no CAS
services, then the pver will have no disks in and will not
be created (because it would be invalid under `conf_pver_width_error`).
In this case, configuration fails to load as the pver is referenced
in the filesystem object. So there is currently no way to create
a configuration with no CAS services.

Waiting for response from DSR team (they may be on holiday).