[HALON-60] Debug info in ignition
We now report how ignition proceeds throughout the given TSs. RS start
is made asynchronous and any node failures are reported but ignition is
allowed to proceed on other nodes. This is a change of behaviour from
before where ignition would simply hang forever in this case (except in
case of a small and unimportant race).