HALON-601: Port non-distributed tests to H.run
The main idea is that we update the runner to be very careful about not
actually using mero; this is OK for most part as halon:m0d mock is used
to replaced the service and most tests don't rely on other
functionality. Beyond that, `syncToBS` and conf validation were being
invoked: we deal with this by inserting appropriate `HalonVars` before
the test runs to instruct RC to not perform mero actions. The only test
relying on the actual, real result from mero is conf validation test: we
leave this test with `[require-mero]` and simply don't validate in other cases.