[HALON-546] On reconnect, reconnect halon:m0d
In case of a network breakage, we lose connection with the node and any
connections on it but the service and processes stay alive. This means
that when the node comes back up (without reboot), we won't be able to
communicate with it straight away: we have to either reconnect to old
connections or establish new ones.

Reconnecting is difficult: we have to wait for old connections to drop
and account for every connection. Instead, we instruct halon:m0d service
to establish new connections (confusingly called "reconnect") restarting
ha_interface and helper processes that are used to communicate with the
service.