HALON-615: Use slots for disk access
We remove the `Enclosure -> StorageDevice` relation. For most part, this
relation is populated during loading of initial data where this
information is now available.

An odd case is metadata/RAID drives which are not in facts file.
Fortunately, drive HPI message should come before RAID message and is
able to create and connect the drive to a slot and then to an enclosure.

The relevant tests were updated to compensate for this change: HPI
message is sent where appropriate and changes in drive replacement tests
were made to account for slots.