HALON-648: RC<->Service communication interfaces
Implement a form of RFC021.

Main changes is that all potential cross-node communication between a
service and RC now goes through some specified interface. Interfaces are
per-service. Type-wise it is akin to using a typed channel in that we
have a single type that we listen for and single (different) type that
we send. However unlike channels, we do not persist this medium of
communication and we are our own middle-man: this allows for
implementation of versioning, message migrations, changes in protocol
&c. For same-node/internal communications, `usend` and friends are still
OK: these are guaranteed to be same versions. Indeed some services
merely decode the message from the interface and pass it on with `usend`
to some internal process.

On top of the core changes, I have made many types used strict when
touching the code around them. There is no good reason to keep these
lazy.

SSPL services and Mero service no longer store their channels in the
graph. This means we do not have to worry about these channels becoming
stale nor have to wait for these to be initialised in RG separately to
the service itself. It simplifies the rules. For Mero service in
particular, we no longer have ~4 different message types indicating
various states that the service could be in. When the associated process
is online, the service is ready to use. A helper retrieving the
interface that performs this check is provided.

These interfaces have some costs. All messages from services go through
a rule in RC which examines them, determines which service's interface
to use to decode the content and packs the message into a format other
rules expect. Rules expecting messages from these services changed as
well: we no longer get the nice type-based message listens in phases. We
have a single type so we have to examine the value to see if it's a
message we're interested in. `setPhaseIf` with extraction functions was
sprinkled around. This is what broke most test as well: notably, these
tests broke silently for the most part because they were still sending
unwrapped types and nothing was listening for these anymore and it was
not a type error. If tests instead use interfaces where they are dealing
with service messages, we get a type error. This was done in few of the
test modules.