CASTOR-2396: Fix pool notification in partial SNS
The problem is that pool is already in M0_NC_REPAIR status in some
scenarios, such as in CASTOR-2396. This is fine but it means
`applyStateChanges` does not generate a state transition and therefore a
mero notification. We receive notification acks about the drives but
never hear about the pool. Notifier mechanism properly times out
while waiting for this.

The proper solution to completely eliminate this class of bug (which is
all over RC and has varying degrees of workarounds in various places)
has been filed under HALON-678. This is just a quick fix in the
meantime.