HALON-683: Versioned service/RC communication
Extend service interfaces to versioned messages. The main idea is that
we always send interface-current version of messages but allow the other
side to reply with their version asking us to re-encode. This happens
when the other side is of lower version. This accomodates for the common
case of everything being same version with additional messages happening
only when versions mismatch.

WireFormat was given a phantom type to denote what sort of thing it's
encoding. This is just a programming aid and may be removed; indeed it
can cause problems: if we're not careful and promulgate `WireFormat
fromSvc` instead of `WireFormat ()`, RC will not be able to handle it.

For messages across versions, see `unsafeResendUnwrapped`. It basically
decodes the raw message (if possible) and sends it to itself with
appropriate fingerprint. It's able to deal with both messages to service
and messages returned from RC.

`HasInterface` moved to TF sparing us from carrying around evidence
along with `Configuration` and forcing all services to define an
`Interface`.

Known bugs
=============
We may leak messages in EQ. Currently the way things work is that
services send a replicated message to RC and RC resends it to itself
with that original UUID. When relevant rules process the message, the
original message gets removed from EQ. This works great as when RC
restarts, the message is re-wrapped and re-sent so everything behaves as
expected.

A rarer scenario is when a message comes to RC from service but RC can
not decode it. In this case RC asks the service to re-encode and resend
but:

* We don't know when (or if) the service receives this request.
* Even if it does, the new message that comes in is not associated with
  previous one so we can't know to acknowledge it.
* Messages to services aren't replicated so we can't rely on replication
  as our "ack" mechanism like `service -> RC` direction does.

Currently we just send a re-encode request to the service and ack the
message straight away. This way we won't leak but we can easily lose
messages (if service doesn't receive/dies/restarts) or duplicate (if RC
sends message to service but before it processes the message, rule
restarts, another request).

Improvements
=============
It can be made easier to compose interfaces of different types. Say an
old message comes for old type: the new interface will read it and use
its decoder. Problems start when it's something like `fromSvc_new` message
returned to the service. There is no nice API for using older interface
to re-encode it and returning `fromSvc_old` of the older interface. Instead
`ifEncodeFromSvc` has to return `WireFormat fromSvc_new` when the
payload is actually WireFormat `fromSvc_old`. It works because in the
end we send `WireFormat ()` to RC but is ugly.

Another improvement is to be more firm about version checking on
messages. Currently we reject messages with higher version than the
interface. Services do this through receiveSvc &c. helpers but RC has to
do it directly. Giving `Interface` a smart constructor and making the
user use accessors which can perform this check would work much more
nicely.

`sendSvcPid` is going to be removed when Frontier service is going to be
removed. There is no reason why this should be a service and is our only
special case. No care was given to this function in this commit.

We can drop `Maybe` around `WireFormat` in encoders: none of the ones we
have are ever expected to fail. It can be re-added in future easily if
we find need for it.

We have to create an interface for service-internal
messages (`ServiceExit` &c). It needs doing but will be in separate
ticket.

`unsafeResendUnwrapped` can in worst case perform full 4 decodes (guard
`toSvc` fails, guard `fromSvc` OK, body `toSvc` fails, body `fromSvc`
OK) on a valid message. It can be cut short if we partially parse
`WireFormat` and examine `wfReceiverVersion` to check determine if it's
`fromSvc` or `toSvc` and cut it to 2 decodes at worst (1 guard, 1 body).

Tests
=============
There are tests for main logic. Actually receiving real messages with
"bad" fingerprints from another version was not tested. To test
something like this we might need to use d-p primitives to create and
send a `Message` with bogus fingerprint. It's a hassle with
d-p-scheduler but should be doable if d-p exposes the right guts (sendCtrlMsg).