HALON-674: process interface requests during stop.
During ha_interface_stop interface can issue events.
It will be waiting for messages delivery after that. But the
thread that processes them is blocked on ha_interface_stop, in
order to avoid this deadlock we are creating new thread that
will be processing requests and kill it as soon as ha_interface_stop
finishes.