CASTOR-2443: Drop service reconnect.
After removal of channels, the only purpose reconnect served was to
trigger STARTED event from mero. However it also turned out to be
harmful, causing a race. Instead of dealing with reconnect and the logic
there, just ask the service to send this event itself. If it's ready and
alive, it will send a message to RC and RC will be none-the-wiser that
it wasn't mero that sent the message.

Note that I've seen issues in case of dropped network connection but am
unable to determine in short-term if it's because of my testing method
or not. Rebooting a node however lets us recover without exhibiting
CASTOR-2443 which is currently a blocker. I therefore suggest merging
this ASAP and testing and fixing (if needed) network drop afterwards.