HALON-702: Clean up halon:sspl behaviour + consume
The service was duplicating any messages it sent to SSPL back onto the
queue that it was listening on. Worse, many tests built on top of this
behaviour.

Fix the problem by using separate exchanges binding to their own queues.
No new queues are created so everything should work without changes on
SSPL side.

Change RMQ proxy to work as a shim between tests and halon:SSPL.
Basically it pretends to be SSPL of sort.
Single-queue-multiple-consumers problem is solved by creating
proxy-internal queue that tests send to and messages are forwarded from
that onto the real queue halon:sspl receives from. This works OK because
all our exchanges are topic exchanges with 1:1 relationship to the
queues without any filtering (i.e. basically `direct` exchanges). In
other scenarios we might have had to worry whether a message was
forwarded or not at all as to not mis-report in the tests.

Add implementation into halon:sspl `teardown` service function. The
behaviour is similar to before but we wait for the internal processes to
die instead of instantly reporting success. This exposes somewhat noisy
logs of form

```
    Rebalancing drive goes back to rebalance after reset:               [Service:SSPL] Exception on channel: ConnectionClosedException "closed by user"
[SSPL-HL]: exception ConnectionClosedException "ConnectionClosedException \"closed by user\"" (src/lib/HA/Services/SSPL.hs:210:40 in mero-halon-0.1-3LmulxqZ2GHKfk7qr2okf9:HA.Services.SSPL)
```

This means we're not shutting down gently enough and should revisit this
code. I'm leaving it as-is for now (HALON-707).

Don't patch `setPhaseIfConsume` because we need to change API somehow.
Currently it's broken but the "trivial" fix of looping doesn't work.
Consider

```
setPhaseIfConsume foo …
setPhaseIfConsume bar …

directly rule_init $ do
  switch [foo, bar]
```

With "trivial" fix, `bar` never fires. Yes, it does sound like job for
dispatcher, again.