HALON-710: Clean SSPL teardown/restart.
Remove seeming random MVar locking, flatten out the service a bit.
Switch directly to `TChan` and leverage `orElse` with `TVar`s for work
processing and finalisation signalling.

We now die cleanly. Move to only catching AMQPExceptions and let any
other exceptions bubble up and kill the service if needed. If we wanted
to, for simplicity we could reduce concurrency and perform RMQ calls in
single worker process. For now synchronisation is hidden in a helper.

Also update rabbitmq perms script to "canonical" values.