HALON-714: Update MM with new resources.
MM will contain the old state and it has to be explicitly told which
resources to remove. The approach taken is to remove all the resources
it knows about and replace them with the set of resources we care about.
This will automatically remove things we can't decode and update binary
representation for things that require it (that we could still decode).