HALON-716: Correctly handle notification timeout
Currently if notifications to Mero are neither cancelled (due to
process being marked as FAILED) nor acknowledged by Mero, Halon will:

- Wait indefinitely for the notification to be acknowledged.
- Fail in other places which are waiting for an object to change state.

This can result in spurious behaviour, such as one process failing
due to another process not acknowledging a message.

This PR introduces a timeout on notification acknowledgement. When this
timeout expires, those Mero processes which fail to acknowledge will
be marked as failed, and the rest of the system should carry on as
usual.