rfc/025-multiple-pools.md: Add field prefixes
Short and simple identifiers (e.g., `pools`) are suitable for definitions
with small scope of visibility, but not quite so for functions that are used
in several modules.  Grepping for a simple identifier is likely to generate
a lot of noise.

We could have allowed fields without prefixes in halon_facts.yaml
while still having prefixes in Halon code, but decided not to do so
for the following reasons:
0. Having different identifiers in halon_facts.yaml and in Haskell code
   would be a source of confusion.
1. Some of existing identifiers in halon_facts.yaml already have
   prefixes (e.g., `id_racks`, `enc_hosts`).
2. This would complicate YAML parsing code
   (see `halonConfigOptions` in
   mero-halon/src/lib/HA/Resources/Castor/Initial.hs).

+ Fix misleading sentence: _all_ disks are being assigned to the IO pool,
  while some of them are _additionally_ assigned to the MD pool.

[skip ci]