scripts: add m0provision
This is a helper script for running Halon and Mero in a development
environment on multiple machines. It provides several commands that
abstract common routine tasks which have to be performed in order to
test Mero and Halon together when they are built in their source
directories.

Assuming that a user has already built Mero and Halon, for example by
running `m0 make` and `h0 make`, they can now start Halon and Mero on
their local cluster with the help of `m0provision` script. The following
example demonstrates a typical sequence of commands:

  # this has to be done only once, after a cluster is created
  m0provision scankeys cmu.local,ssu1.local,ssu2.local,client1.local

  # bring Halon services up on all hosts
  m0provision up -c cluster.yml cmu.local,ssu1.local,ssu2.local,client1.local

  # do usual Halon stuff
  halonctl mero bootstrap
  halonctl mero status
  # --> run some tests here <--
  halonctl mero stop

  # completely remove Halon+Mero meta-data and services on all hosts
  m0provision destroy cmu.local,ssu1.local,ssu2.local,client1.local