rpm: enable override of the default rpmbuild output directory
By default `make rpm-dev` command sets up a working directory for
`rpmbuild` inside top-level Halon source tree directory. Usually it
doesn't matter, but when Halon sources are on a filesystem like NFS or
HGFS (VMware shared directory) it can be a problem due to file ownership
and permission requirements - `rpmbuild` reports "Bad owner/group" error
if some of the file permissions don't match current user/group ID. For
example if "spec" file belongs to a different user or it's access
permissions are too lax. In such situation the most convenient way of
satisfying `rpmbuild` requirements is to create it's working directory
on a local filesystem outside of Halon source tree, which now can be
done by setting makefile variable on a command-line:

    make rpm-dev RPMBUILD_DIR=~/rpmbuild