rpm: place rpms in ~/rpmbuild by default
Historically Halon used to store rpms in the rpmbuild/ directory at the
top level of its source tree. To unify Mero and Halon build
configuration the default location is changed to be the same as used by
Mero. The old behaviour can be achived via setting the RPMBUILD_DIR
environment variable, like this:

  make rpm-dev RPMBUILD_DIR=$PWD/rpmbuild