Import Mero.Spiel qualified
Mero.Spiel module is exposed by `confc` package. Import this module
either qualified or with an explicit import list, in order to see
which definitions of `confc` package are actually used.

We intend to shrink interfaces of `confc` package by removing unused code.
For this it is important to know what *is* used.