Import Mero.ConfC with explicit import list
Mero.ConfC is exposed by `confc` package. This package contains unused
code, which we are going to delete. (This code is incompatible with new
Mero conf schema. We have to either update it or delete it. Deletion
is simpler.)

The purpose of this change is to clarify which definitions of `confc`
package are used.