scripts/h0: Fix word splitting issue
Due to the bug in initial implementation, `h0 test -p 'stop then start'`
used to be recognized as a sequence of `h0 test -p`, `h0 stop then`, and
`h0 start` commands.  This patch fixes this problem; now
`'stop then start'` is properly treated as an argument of `-p` option
and is not split into separate words.