Mero/Rules.hs: fix s3server startup on mero stop/start
Apparently, this case was never tested before. s3server just
never started anymore on cluster stop/start (after initial
bootstrap) because of this silly bug in ruleDixInit code
for the case when the DIX is already initialised.

Each case in ruleDixInit rule has to be finished either with
`switch' or `continue' (see the context code in the patch).
The "DIX subsystem already initialised" case was not finished
with either of them, so it just got stuck there after printing
the debug message in the desicion log. The one line fix is just
`continue finish' added at the end of the case.

Reviewed-by: Valery V. Vorotyntsev <valery.vorotyntsev@seagate.com>