mero-halon: getUnstartedProcesses() fixed
It would not return the process which has CST_CAS or
CST_ADDB2 or CST_SNS_REP services (any not from the list),
for example.

Now we refactored the code to get rid of the ill-heuristics
to determine the type of the process (client/non-client)
by the list of its services. Instead, we explicitly check
its ProcessType.