HALON-797: UT: fix sendProcessEvents:guessProcessType()
It could guess wrong process type in a cases, when CST_CAS
service or CST_ADDB2 service (or any not present in
the comparison list) was provided in the argument list.
In fact, it did set TAG_M0_CONF_HA_PROCESS_KERNEL type
for all the TAG_M0_CONF_HA_PROCESS_M0D type cases in the
stop-start UT.

Now we fixed the algorithm to specify in the matching list
only the services which are possible for server (M0D)
processes only. And if any of those services is present
in the argument list - it means we are dealing with M0D.
The algorithm will still work in case when some new unknown
services will be introduced in future to M0D (given that
it would not be the single one configured for M0D process).

The patch to the UT detected the race problem with services
offline notifications ACKs and ruleProcessStop. The latter
does not wait for these ACKs and proceeds with stopping.
After  stop completes the stop-start UT continues with
starting, but the stale ACKs would still come and affect
the new services states already. As result, the UT did not
pass.

We fixed this by not waiting for the offline notification
ACKs for services in the notification-handler.  This also
might fix a subtle race cases with cluster stop command on
a big cluster configurations when some ACKs may not come
from already stopped processes in which case the command
will fail with a timeout.