cep/phase: list to sequence optimisation
Adding at the end of the list is O(n) expensive, so we
optimise this part at runPhaseM by changing the list
of phases to the sequence (Data.Sequence). This gives
us about 4% boost in performance on 60 client processes
startup (96.18 secs vs. 100.66 secs total CPU time
according to profile results) and might give more on
a larger configurations.