HALON-830: fix intermittent fork-remove-messages UT failures
In this test 3 SM instances are started one after another
through their phases.  Sometimes the next SM instance
starts working before the previous one is finished yet,
causing the resulting output to be overlapped. This effect
seems to be harmless, because within the single SM instance
the messages are still handled in right order.

So we just fix the test by excluding the check for the
1st SM phase output.