HALON-849: fix checkDiskFailureWithinTolerance
checkDiskFailureWithinTolerance could return false-positive
result on multiple pools. If some failed disk belongs to
different pools with different Ks, it would check the pool
which has maximum K only. But it may well be that the
failed disk is not tolerable already in another pool(s)
with lesser K.

Now we check the pool with has minimum K in the list.
If we can tolerate disk failure in it - we can tolerate
it in all the others that has the same or bigger K.