HALON-795: fix start/stop on a big clusters
The problem was with the number of notification Acks
we had to process at RC for each notification which is
sent on a new epoch to every single online process in the
cluster. New epoch happens when, for example, some process
change its state. So in N-process cluster configurations
this resulted in O(N^2) notification messages in total to
be sent and Ack-ed on such scenarios as cluster startup,
for example. While sending is relatively cheap, the
processing of Acks is quite expensive (because it goes via
replicated EQ) and Halon struggled to process such amount
of events in a reasonable time.  (2 minutes currently.)

Now we don't send every single notification directly to
the destination process, but accumulate and aggregate them
on SATs for some time (2 secs by default, configurable via
notification-aggr-delay mero vars parameter) into a single
notification packet which is then sent to all processes
underneath these SATs.  Such approach decreases the amount
of notification Acks dramatically, down to O(N).

The patch allowed to start the cluster with 300 configured
processes on a singlenode setup.

Reviewed-by: Valery V. Vorotyntsev <valery.vorotyntsev@seagate.com>