[HALON-882] Serialize fids to JSON as strings
Serialize fids to JSON as strings in "<container>:<key>" format, where
<container> and <key> are hexadecimal numerals (e.g.,
"0x6500000000000001:0x5").

`f_container` part of a fid is a large (> 2^53) integer, which
JavaScript runtime (web browser's JS engine, `jq` utility) cannot
handle properly.

    $ echo $((0x6500000000000001))
    7277816997830721537
    $ echo $((0x6500000000000001)) | jq .
    7277816997830722000

This is not a problem when fid is serialized to JSON as string.
Another benefit is that hexadecimal representation of a fid is more
comprehensible than decimal representation.